added persist-instance-on-dail option to the tests

+++ omnistrate_tests/test_cluster_replicas.py
@@ -51,7 +51,7 @@ parser.add_argument("--aof-config", required=False, default="always")
 parser.add_argument("--host-count", required=False, default="6")
 parser.add_argument("--cluster-replicas", required=False, default="1")
 parser.add_argument("--shards", required=False, default="3")
-
+parser.add_argument("--persist-instance-on-fail",required=False,default=False)
 parser.add_argument("--ensure-mz-distribution", action="store_true")
 
 parser.set_defaults(tls=False)
@@ -67,8 +67,9 @@ def signal_handler(sig, frame):
     sys.exit(0)
 
 
-signal.signal(signal.SIGINT, signal_handler)
-signal.signal(signal.SIGTERM, signal_handler)
+if args.persist_instance_on_fail is False:
+    signal.signal(signal.SIGTERM, signal_handler)
+    signal.signal(signal.SIGINT, signal_handler)
 
 current_host_count = int(args.host_count)
 current_replicas_count = int(args.cluster_replicas)
@@ -157,7 +158,8 @@ def test_cluster_replicas():
 
     except Exception as e:
         logging.exception(e)
-        instance.delete(False)
+        if args.persist_instance_on_fail is False:
+            instance.delete(False)
         raise e
 
     # Delete instance
